[ PWM CONFIG ] : Adds N20 pwm control

+++ firmware/1_N20Control/main/include/tuning_http_server.h
@@ -32,8 +32,7 @@ typedef struct pid_const
     float kp;
     float ki;
     float kd;
-    int setpoint;
-    
+    float setpoint;
     bool val_changed;
 } pid_const_t;
